mergeado de Luis

+++ BolsaScripts/inversionHistoricaSoloFuturoSimplificado.sh
@@ -3,7 +3,11 @@
 #set -e
 
 #################### DESCRIPCIÓN Y PARÁMETROS FUNDAMENTALES ###############################
-# Permite sacar las predicciones en un rango de tiempos del pasado. El modelo predictor NO SERÁ el usado en 
+# Permite sacar las predicciones en un rango de tiempos del pasado. 
+# Se coge rel modelo de la matigüedad máxima (no lo genera este script). Luego se predice el futuro de la antigüedad mínima, y
+# se calcula también el de las antigüedades intermedias en base a lo ya descargado para la antigüedad mínima (fichero COMPLETO.csv), 
+# borrando previamente los días necesarios en le fichero COMPLETO y prediciendo el futuro sin redescargar.
+# El modelo predictor NO SERÁ el usado en 
 # antigüedad =0, sino que cogeremos la antigüedad máxima deseada, y lo entrenaremos con datos anteriores.
 # POR TANTO, LOS RESULTADOS PUEDEN DIFERIR DE LOS QUE SE USEN EN EL FUTURO CON DINERO REAL.
 # Por otro lado, SIEMPRE SE DEBEN ACTIVAR LAS DESCARGAS (en parametros.config) para que, en cada 
@@ -75,7 +79,7 @@ source ${PARAMS_CONFIG}
 
 DIR_BASE="/bolsa/"
 DIR_LOGS="${DIR_BASE}logs/"
-LOG_INVERSION="${DIR_LOGS}inversion.log"
+LOG_INVERSION="${DIR_LOGS}inversionHistoricaSoloFuturoSimplicado.log"
 DIR_FUT_SUBGRUPOS="${DIR_BASE}futuro/subgrupos/"
 DIR_JAVA="${DIR_CODIGOS}BolsaJava/"
 PATH_JAR="${DIR_JAVA}target/bolsajava-1.0-jar-with-dependencies.jar"
@@ -83,21 +87,69 @@ DIR_GITHUB_INVERSION="${DIR_CODIGOS}inversion/"
 
 
 #################################### CÓDIGO ###########################################################
-
+DESCARGAR="INVALIDO"
 for (( ANTIGUEDAD=${ANTIGUEDAD_MINIMA}; ANTIGUEDAD<=${ANTIGUEDAD_MAXIMA}; ANTIGUEDAD++ ))
 do  
 
-	rm -Rf ${DIR_BASE}futuro/ >>${LOG_INVERSION}
-	crearCarpetaSiNoExiste "${DIR_FUT_SUBGRUPOS}"
-	crearCarpetaSiNoExiste "${DIR_FUT_SUBGRUPOS}brutos/"
-	crearCarpetaSiNoExiste "${DIR_FUT_SUBGRUPOS}brutos_csv/"
-	crearCarpetaSiNoExiste "${DIR_FUT_SUBGRUPOS}limpios/"
-	crearCarpetaSiNoExiste "${DIR_FUT_SUBGRUPOS}elaborados/"
 	
+
+######## Para la antigüedad mínima (normalmente=0), se vacía la carpeta de futuro y se activan las descargas, pero para el resto no (porque se reutilizarán)
+	if [ ${ANTIGUEDAD} -eq ${ANTIGUEDAD_MINIMA} ]; then
+
+        	DESCARGAR="S"
+
+		rm -Rf ${DIR_BASE}futuro/ >>${LOG_INVERSION}
+		crearCarpetaSiNoExiste "${DIR_FUT_SUBGRUPOS}"
+		crearCarpetaSiNoExiste "${DIR_FUT_SUBGRUPOS}brutos/"
+		crearCarpetaSiNoExiste "${DIR_FUT_SUBGRUPOS}brutos_csv/"
+		crearCarpetaSiNoExiste "${DIR_FUT_SUBGRUPOS}limpios/"
+		crearCarpetaSiNoExiste "${DIR_FUT_SUBGRUPOS}elaborados/"
+	
+	fi
+
+######## Para antigüedades mayores que la mínima, se asume que el fichero COMPLETO.csv de la antigüedad mínima ya existe, así que no se volverá a descargar, sino que se borrarán los días posteriores a esa antigüedad mínima (en la práctica se eliminan todas las filas de un día, para cada iteración), y se predecirá.
+# El fichero resultante de la predicción sobreescribirá al anterior, porque no tiene fecha en su nombre, sino sólo variará su contenido.
+# Antes que nada, se eliminan pasadas predicciones: ficheros *COMPLETO_PREDICCION*
+	if [ ${ANTIGUEDAD} -gt ${ANTIGUEDAD_MINIMA} ]; then
+                # Se evita descargar COMPLETO.csv
+		DESCARGAR="N"
+
+                # Se eliminan ficheros de anitguas predicciones (en TODOS los subgrupos)
+                while IFS= read -r -d '' -u 9
+		do
+			if [[ $REPLY == *"COMPLETO_PREDICCION"* ]]; then
+				echo "Se elimina el fichero  ${REPLY}  ..."  >>${LOG_INVERSION}
+				ficheronombre=$(basename $REPLY)
+				directorio=$(dirname $REPLY)
+				rm "${directorio}/${ficheronombre}" >> ${LOG_INVERSION}
+			fi
+		done 9< <( find ${DIR_FUT_SUBGRUPOS} -type f -exec printf '%s\0' {} + )
+
+                # Se elimina el siguiente día de COMPLETO.csv (en TODOS los subgrupos)
+		while IFS= read -r -d '' -u 9
+		do
+			if [[ $REPLY == *"COMPLETO.csv" ]]; then
+				echo "Procesamos  ${REPLY}  para eliminar todas las filas del día más reciente ..."  >>${LOG_INVERSION}
+				ficheronombre=$(basename $REPLY)
+				directorio=$(dirname $REPLY)
+				$PYTHON_MOTOR "${PYTHON_SCRIPTS}bolsa/QuitarDiasMasRecientes.py" "${directorio}/${ficheronombre}" "1" "${directorio}/" "${ficheronombre}" >> ${LOG_INVERSION}
+			fi
+		done 9< <( find ${DIR_FUT_SUBGRUPOS} -type f -exec printf '%s\0' {} + )
+
+
+	fi
+
 	echo -e $( date '+%Y%m%d_%H%M%S' )" Ejecución del futuro (para velas de antiguedad=${ANTIGUEDAD}) con TODAS LAS EMPRESAS (lista DIRECTA ó INVERSA, ya da igual, no estamos mirando overfitting)..." >>${LOG_INVERSION}
-	MIN_COBERTURA_CLUSTER=0    # Para predecir, cojo lo que haya, sin minimos. El modelo ya lo hemos entrenado
-	MIN_EMPRESAS_POR_CLUSTER=1   # Para predecir, cojo lo que haya, sin minimos. El modelo ya lo hemos entrenado
-	${PATH_SCRIPTS}master.sh "futuro" "${ANTIGUEDAD}" "0" "${ACTIVAR_DESCARGAS}" "S" "${S}" "${X}" "${R}" "${M}" "${F}" "${B}" "${NUM_EMPRESAS_INVERSION}" "${UMBRAL_SUBIDA_POR_VELA}" "${UMBRAL_MINIMO_GRAN_VELA}" "${MIN_COBERTURA_CLUSTER}" "${MIN_EMPRESAS_POR_CLUSTER}" "20001111" "20991111" "${MAX_NUM_FEAT_REDUCIDAS}" "${CAPA5_MAX_FILAS_ENTRADA}" "${DINAMICA1}" "${DINAMICA2}" 2>>${LOG_INVERSION} 1>>${LOG_INVERSION}
+	MIN_COBERTURA_CLUSTER=0    # Para predecir, cojo lo que haya, sin mínimos. El modelo ya lo hemos entrenado
+	MIN_EMPRESAS_POR_CLUSTER=1   # Para predecir, cojo lo que haya, sin mínimos. El modelo ya lo hemos entrenado
+
+        if [ ${ANTIGUEDAD} -eq ${ANTIGUEDAD_MINIMA} ]; then
+	        ${PATH_SCRIPTS}master.sh "futuro" "${ANTIGUEDAD}" "0" "${DESCARGAR}" "S" "${S}" "${X}" "${R}" "${M}" "${F}" "${B}" "${NUM_EMPRESAS_INVERSION}" "${UMBRAL_SUBIDA_POR_VELA}" "${UMBRAL_MINIMO_GRAN_VELA}" "${MIN_COBERTURA_CLUSTER}" "${MIN_EMPRESAS_POR_CLUSTER}" "20001111" "20991111" "${MAX_NUM_FEAT_REDUCIDAS}" "${CAPA5_MAX_FILAS_ENTRADA}" "${DINAMICA1}" "${DINAMICA2}" 2>>${LOG_INVERSION} 1>>${LOG_INVERSION}
+        fi
+
+        if [ ${ANTIGUEDAD} -gt ${ANTIGUEDAD_MINIMA} ]; then
+        	${PATH_SCRIPTS}masterSimplificado.sh "futuro" "${ANTIGUEDAD}" "0" "${DESCARGAR}" "S" "${S}" "${X}" "${R}" "${M}" "${F}" "${B}" "${NUM_EMPRESAS_INVERSION}" "${UMBRAL_SUBIDA_POR_VELA}" "${UMBRAL_MINIMO_GRAN_VELA}" "${MIN_COBERTURA_CLUSTER}" "${MIN_EMPRESAS_POR_CLUSTER}" "20001111" "20991111" "${MAX_NUM_FEAT_REDUCIDAS}" "${CAPA5_MAX_FILAS_ENTRADA}" "${DINAMICA1}" "${DINAMICA2}" 2>>${LOG_INVERSION} 1>>${LOG_INVERSION}
+        fi
 	
 	while IFS= read -r -d '' -u 9
 	do
@@ -108,6 +160,8 @@ do
 			$PYTHON_MOTOR "${PYTHON_SCRIPTS}bolsa/InversionUtils.py" "${directorio}/${ficheronombre}" "0" "${DIR_DROPBOX}" "${ficheronombre}" >> ${LOG_INVERSION}
 		fi
 	done 9< <( find ${DIR_FUT_SUBGRUPOS} -type f -exec printf '%s\0' {} + )
+
+
 
 done
